jig/collector/jig_ast配下のファイルをjig/collector/domain/ast配下に移動
