+++ jig/collector/domain/__init__.py
@@ -2,8 +2,7 @@ import dataclasses
 import os
 from typing import List, Tuple, Optional
 
-from jig.collector.jig_ast import JigSourceCode
-from jig.collector.domain.ast import ClassDef
+from jig.collector.domain.ast import ClassDef, JigSourceCode
 from jig.collector.domain.ast import Import, ImportFrom
