+++ jig/collector/domain/ast/__init__.py
@@ -48,3 +48,90 @@ class ImportFrom:
             level=import_from.level,
             _ast=import_from,
         )
+
+
+@dataclasses.dataclass(frozen=True)
+class JigAST:
+    _ast: ast.AST
+
+    @classmethod
+    def parse(cls, source: str, filename: str = "<unknown>") -> "JigAST":
+        tree = ast.parse(source=source, filename=filename)
+        return cls(tree)
+
+    def imports(self) -> List[Import]:
+        visitor = ImportVisitor()
+        visitor.visit(self._ast)
+
+        return visitor.imports
+
+    def import_froms(self) -> List[ImportFrom]:
+        visitor = ImportFromVisitor()
+        visitor.visit(self._ast)
+
+        return visitor.import_froms
+
+    def class_defs(self) -> List[ClassDef]:
+        visitor = ClassDefVisitor()
+        visitor.visit(self._ast)
+
+        return visitor.class_defs
+
+
+@dataclasses.dataclass(frozen=True)
+class JigSourceCode:
+    imports: List[Import]
+    import_froms: List[ImportFrom]
+    class_defs: List[ClassDef]
+
+    @classmethod
+    def build(cls, source: str, filename: str) -> "JigSourceCode":
+        jig_ast = JigAST.parse(source=source, filename=filename)
+
+        return cls(
+            imports=jig_ast.imports(),
+            import_froms=jig_ast.import_froms(),
+            class_defs=jig_ast.class_defs(),
+        )
+
+
+@dataclasses.dataclass
+class ClassDefVisitor(ast.NodeVisitor):
+    class_defs: List[ClassDef] = dataclasses.field(default_factory=list)
+
+    def visit_ClassDef(self, node):
+        self.class_defs.append(ClassDef(name=node.name, _ast=node,))
+
+
+@dataclasses.dataclass
+class ImportVisitor(ast.NodeVisitor):
+    imports: List[Import] = dataclasses.field(default_factory=list)
+
+    def visit_Import(self, node):
+        self.imports.append(
+            Import(
+                names=[
+                    Alias(name=name_alias.name, asname=name_alias.asname)
+                    for name_alias in node.names
+                ],
+                _ast=node,
+            )
+        )
+
+
+@dataclasses.dataclass
+class ImportFromVisitor(ast.NodeVisitor):
+    import_froms: List[ImportFrom] = dataclasses.field(default_factory=list)
+
+    def visit_ImportFrom(self, node):
+        self.import_froms.append(
+            ImportFrom(
+                module=node.module,
+                names=[
+                    Alias(name=name_alias.name, asname=name_alias.asname)
+                    for name_alias in node.names
+                ],
+                level=node.level,
+                _ast=node,
+            )
+        )
