+++ tests/collector/jig_ast/test_jig_ast.py
@@ -1,5 +1,4 @@
-from jig.collector.domain.ast import Import, ImportFrom
-from jig.collector.jig_ast import JigAST
+from jig.collector.domain.ast import Import, ImportFrom, JigAST
 
 
 class TestJigASTImport:
